feat(#537): move InnerClass attribute upper
